Move Parser to GUID
GUID must be independent of the rest of the library. Additionally,
Parser code is only used by GUID.

Although some consider nested classes to be a smell, this is not true in
all cases. If you think that's the case, I hope you can forgive this
one.
